+++ tests/testthat/test-map_ontology_terms.R
@@ -27,7 +27,7 @@ test_that("map_ontology_terms works", {
   
   #### CL ####
   cl <-  get_ontology("cl", )
-  testthat::expect_lte(nrow(subset(cl@elementMetadata, is.na(name))), 3)
+  testthat::expect_lte(nrow(subset(cl@elementMetadata, is.na(name))), 20)
   terms <- cl@terms
   run_tests(ont = ont, terms = terms) 
 })